refactor(ai-service): split imitate_expert — extract TrackExpertModel + ExpertPositionLearner
Second slice of the Page-5 ML splits. Pulls the two pure-model classes
out of the imitation service into their own module; the broader
``ExpertImitateLearningService`` orchestrator (training entry point,
backend serialisation, prediction wrapper) stays in service.py.

Extracted:
  app/ml/imitation/service.py
    ├─ _format_debug_message helper
    ├─ TrackExpertModel       (per-track interpolation model)
    └─ ExpertPositionLearner  (multi-track manager owning TrackExpertModel dict)
                              → app/ml/imitation/model.py

model.py is a pure leaf: imports numpy/pandas/logging + app.domain
only. No I/O, no service-layer code. ExpertImitateLearningService
imports the two classes back at the top of service.py so all existing
internal references and external callers see the same surface.

Validated:
  • TrackExpertModel + ExpertPositionLearner identity preserved via
    both `app.ml.imitation.model` and `app.ml.imitation.service`
  • Smoke construction works (TrackExpertModel("silverstone"),
    ExpertPositionLearner())
  • lint-imports 11 kept, 0 broken
  • pytest 24 passed
  • /health 200
